Group 'kopia server' commands in single file
